Extend timeout of ui.DragTabInClamshellPerf
When I tried manually, I found there's a risk of timeout on this
test because of the RunMultiple. Setting 3 minutes would mitigate
the problem.

This is not happening on the lab but many of the runs took ~1m30s,
so it would be safer to increase the timeout.

Also this CL reduces the duration of drag action; that would make
the test slightly faster (considering it runs 11 times now).

BUG=none
TEST=tast run ui.DragTabInClamshellPerf on krane

Change-Id: I5a9be0e8e3e91b030b595740f454ebbf77d6a157