Pass ctx to metrics.Run and its variations
metrics.Run (and RunAndWaitAll etc) has been getting a function
without context, assuming Run's first argument can be used.

Recently this is not going to be the case; some utility functions
(like perfutil.RunMultiple) can tweak the context's deadline and
then invoke metrics.RunAndWaitAll. In this case, the inner function
should use the tweaked context.

With this CL, metrics.Run passes the given context to the inner
function, and that way the inner function recognizes the tweaeked
context, finishes with the right deadline.

BUG=none
TEST=fast_build.sh -C

Change-Id: I219544bc4654bda51f71ae6424b6618d623d3224