security.SELinuxAuditBasic: Add retries to read audit.log
tast.security.SELinuxAuditBasic sometimes fails because this test
reads audit.log to make sure that there is a log entry of selinux
in the log file, but there is a posssibility of delay in flushing
the file and the target log is not written yet.

This CL adds retries to read logs in case of delay.

This bug has been there for a while but recently revailed by
introducing recent croslog CLs, since it changed the timing of
this test.

BUG=chromium:1102718
TEST=Ran CQ

Change-Id: I24c0b64b92f2fc95aaf7560c5d4ba15d184d19d9