crostini: use the UI for the restart test
The current restart test directly restarts crostini via dbus.  This
means that all of the management code in the UI gets skipped, and is
unaware that crostini restarted.  This leads to a number of problems,
including failing tests that rely on state set up by the UI code, such
as the SharedFontFiles test.

This change shuts down and restarts crostini in the same way a user
would do, fixing the failing tests and increasing test coverage.

BUG=chromium:1095945,chromium:1101211
TEST=Ran restart test with 30 iterations, ran restart and
sharedfontfiles in previously broken download configuration

Change-Id: Ifdeeb1ac726b7e1abd34a85a61750ec2b988a83b