arc.CacheValidation: fix bad regexp match and hard-coding of the build ID.
Without this change, the code unconditionally fails on my Eve dut. With
the change, the test has passed at least three times in a row.

Note that the test fails to compile at ToT for some reason that I haven't
yet figured out (unrelated to my change); it passes on m/stable.

BUG=b:162939498
TEST=tast -verbose run  -build=true $DUT_IP arc.CacheValidation.pi_container

Change-Id: I7771fe07ba7950cedf95237e313c4f818c1c87e5