Fix ui.SplitViewResizePerf failures on atlas
The test sometimes fails because it fails to wait for the metrics
for the divider animation. During my debug, I noticed that this
is due to a side effect of WaitForLocationChangesCompleted.
My guess is that the communication between the test extension and
the browser process can be stuck due to the communication of
accessibility events, which is not wanted at all.

Actually, WaitForLocationChangesCompleted wouldn't be necessary
at all for this case, because RunAndWaitAll should wait for the
metrics of the divider, which should happen after the animation
finishes. Thus, this CL simply removes the related part of the
code, and adds some comments of why it is okay.

BUG=chromium:1108029
TEST=tast run ui.SplitViewResizePerf on atlas

Change-Id: I9fcb8e85cd81626122894a5eb830d77d038d2ff2