Fix ash.EnterShelfOverflow
It seems that IsOverflow flag is not quite reliable; it may
return an intermediate state, so IsOverflow can turn into false
very quickly in such case.

Instead, this CL checks the icon bounds and compare the bounds of
the display; it should be overflowing when at least one of the
icon bounds is outside of the display.

BUG=chromium:1105619
TEST=tast run ui.HotseatScrollPerf.clamshell_mode on nocturne

Change-Id: Ic3c9b1cc0bff450705aa757195a42795f8be9b20