tast-tests: hwsec: Add hwsec.DictionaryAttackLockoutResetTPM1 test
hwsec.DictionaryAttackLockoutResetTPM1 test is similar to
hwsec.DictionaryAttackLockoutReset local test, except that this test is
specifically for TPMv1.2, while the original local test is specifically
for TPMv2.0. 2 versions of the same test is required because the way
different TPM version deal with NVRAM access and the available NVRAM space.

Additionally, ReadSpace in tpm_manager utility is added because
it's used by this test.

BUG=b:160260730
TEST=tast run $dut hwsec.DictionaryAttackLockoutResetTPM1 on relm

Change-Id: Ie59ab22f8e1432146203090cf5a6de725fad8039
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2299783
Commit-Queue: John L Chen <zuan@chromium.org>
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>