wifi: Add MultiAuth test
This is a port of network_WiFi_MultiAuth from autotest.

This test launches two hostapd instances on the same PHY, with
identical SSID/Channel/Mode but different security settings,
and then verifies that the client can connect to both.

Porting this test required a couple of changes to the router
configuration code, to permit reusing the same PHY and to
ensure that distinct interfaces on the same PHY have unique
MAC addresses (see site_linux_system.py:595 for the Python
equivalent of this behavior).

BUG=b:162017454
TEST=tast run $DUT wifi.MultiAuth

Change-Id: Ie28a882b53487909a5d3842c1cc14aeab73e341d