tast-tests: hwsec: Add take ownership to AttestationEID
Enrollment ID depends on endorsement key, which can only be read when
TPM is ready on TPMv1.2. This change add the take ownership operation
to this test to ensure that the TPM is ready every time the test is run.
This prevent the test from failing incorrectly when TPM is unowned.

BUG=none
TEST=tast run $dut hwsec.AttestationEID

Change-Id: Ic35c9e6aa3574d1fbc3460b7e22e0b49f73d8093