tast-tests: hwsec: Add take ownership to AttestationEID
Enrollment ID depends on endorsement key, which can only be read when
TPM is ready on TPMv1.2. This change add the take ownership operation
to this test to ensure that the TPM is ready every time the test is run.
This prevent the test from failing incorrectly when TPM is unowned.

BUG=none
TEST=tast run $dut hwsec.AttestationEID

Change-Id: Ic35c9e6aa3574d1fbc3460b7e22e0b49f73d8093
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2315470
Commit-Queue: John L Chen <zuan@chromium.org>
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>