tast-tests: wificell: Remove curServicePath field in TestFixture.
As we're adding roaming tests, it is possible that curServicePath on
the object is not the same as SelectedService in shill. Drop the field
and try to get the latest information from shill. For tests which
expect to disconnect from a specific service, the gRPC interface is
also compatible for their usage.

Also add options for ease of cleaning profile when disconnecting.

BUG=None
TEST=tast run $DUT wifi.*

Change-Id: Idd9bd36b900367f4171dbc2337701ef6d47d1fc3