Create firmware.Helper with RPC connection
firmware.Helper will be useful for keeping track of more than just the
RPC; see go/tast-fw-helper for more design context.
This CL gets things started with NewHelper(), h.Close(), and
h.RequireRPCClient() / h.RequireRPCUtils().

BlockingSync is an apt test for starting to use the Helper, because all
it requires is the existence of an RPCUtils. It exercises all Helper
methods added in this CL, and demonstrates how the Helper simplifies
tests.

BUG=b:162068479
TEST=firmware.BlockingSync

Change-Id: I2bfb7d6d4e7ff46118888b1ef4d3a6caa27387fb