Disable ui.ChromeCrashLoggedIn.gpu_process_breakpad
This informational test has been failing 100% of the time for several
weeks. I'm not going to have the bandwidth to look at it for a week or
two. In keeping with the Green Stainless project, disabling the test
until I have time to investigate the failure.

BUG=chromium:1110025
TEST=tast ui.ChromeCrashLoggedIn.*

Change-Id: Id7867ca9a645d977f3db2df60a219eda5ff2d266
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2321136
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>