firmware.Helper: Add RequireConfig,RequirePlatform
With this CL, a test can declare that it needs a firmware.Config with
h.RequireConfig(ctx). The firmware.Config test exercises this
functionality.

The new h.RequireConfig relies on another new method, h.RequirePlatform,
which fetches the DUT's board and model from the firmware utils Platform
RPC. The results are cached so that tests don't need to run the RPC
multiple times.

BUG=b:162068479
TEST=firmware.Config

Change-Id: Ife685527546ef2f59ea1b94789846c7083cb8d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2321139
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>