firmware.Helper: Add RequireConfig,RequirePlatform
With this CL, a test can declare that it needs a firmware.Config with
h.RequireConfig(ctx). The firmware.Config test exercises this
functionality.

The new h.RequireConfig relies on another new method, h.RequirePlatform,
which fetches the DUT's board and model from the firmware utils Platform
RPC. The results are cached so that tests don't need to run the RPC
multiple times.

BUG=b:162068479
TEST=firmware.Config

Change-Id: Ife685527546ef2f59ea1b94789846c7083cb8d89