Extend the timeout of RunAndWait* for UI perftests
Some tests are still failing after running repeatedly, with
"not all histogram changed" error. As I checked the behavior
of such a test, ui.SplitViewResize, and noticed that sometimes
there's a huge delay in ui::ThroughputTracker to report a
number, which flakes the test.

As far as I tried, extending to 1 minute is good enough;
with my 5 consecutive runs on eve (each run conducts 33 times
of resizing operations), the longest delay to report is ~40
seconds.

This CL also updates the error message of histogram so that
the developers can see which metrics are missing.

BUG=chromium:1108029
TEST=tast run ui.SplitViewResize on eve for 5 times

Change-Id: I4a7982157a9e06aa9cc888eeb7e9825897e81da5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2321732
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>