tast-tests: Make cvconn be chrome.TestConn to allow arc a11y tree to be dumped
This CL wraps cvconn instead of chrome.TestConn, for it to be allowed to
be passed to DumpUITreeOnErrorToFile(). Which will dump the full a11y
tree of the application, and will assist with debugging.

This solution is simpler than changing DumpUITreeOnErrorToFile,
which is shown in crrev/c/2311978.
cvconn needs to be passed to DumpUITreeOnErrorToFile, as the changes to
the a11ytree made by cvconn, are not reflected inside of the test
extension connection.

BUG=b:161864703
TEST=make test fail, and confirm dumped log.

Change-Id: I2bad48f7b6e57364a433a752936ffa97a066521f