tast-tests: Disable video.DecodeAccelSanity.vp9_2 on zork
On zork, with the legacy video decoder, vp9 profile 2 puts the GPU into
a bad state. When video.DecodeAccelSanity.vp9_2 runs as part of a suite,
any subsequent tests requiring the GPU (e.g. any that launch Chrome)
will fail.

Disable video.DecodeAccelSanity.vp9_2 on zork until this test switches
to the new video decoder, which does not have this issue on zork.

BUG=b:161878038
TEST=informational-chrome suite tests do not all fail to launch chrome
     after video.DecodeAccelSanity.vp9_2 runs on a trembyle

Change-Id: I96de59a83028edeaa63cbacd7dc77a861bfa56ff