crash: Allow 4 reboots during remote crash tests.
This is a speculative fix to flakiness in early crash test - we
sometimes see the mock-consent and crash-test-in-progress files missing
after the machine comes back from a reboot. So far, it's proven
difficult to figure out why. Having too many reboots is a simple
explanation, and not outside the realm of plausibility.

BUG=chromium:1039378
TEST=tast run $DUT crash.EarlyCrash.*

Change-Id: I205f8a054b500daa548683d981745db5d91d3762
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2325010
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>