Add NC07 Tast-test for WM ARC++ R
Implement NC07: non-resizable/clamshell: immerse via API from
maximized. In NC07 a non-resizable activity is started and window
bounds are stored. Then the immersive button is clicked. This should
take the activity to fullscreen mode. Then the normal button is
clicked, this should take the activity to maximize mode and the
window bounds should be equal to the original window bounds in root.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableClamshell

Change-Id: I243e166d311823f7e5565b729b63259fb9b452e1