Add NT07 Tast-test for WM ARC++ R
Implement NT07: non-resizable/clamshell: immerse via API from
maximized. In NC07 a non-resizable activity is started at tablet mode
and window bounds are stored. Then the immersive button is clicked.
This should take the activity to fullscreen mode. Then the normal
button is clicked, this should take the activity to maximize mode and
the window bounds should be equal to the original window bounds in
root.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableTablet

Change-Id: Ie50ff866e16fec0891e222241a807ca9dba2c694