Stabilize PIP Expand Shelf Icon test in R
This change improves the existing shelf icon expand test to make
it compatible with R. Previously, the test had the assumption that
the restore window state would be Maximized, which is not true in
all cases. It now explicitly sets the initial window states, then
enters PIP and then verifies restoring returns it to the initial
window state.

BUG=b:161834503
TEST=tast run arc.PIP.vm

Change-Id: I22eec01dbbaec508d90790f67564174568f98130
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2327874
Tested-by: Jorge Gil <jorgegil@google.com>
Commit-Queue: Jorge Gil <jorgegil@google.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Kazuki Takise <takise@chromium.org>