EthernetStaticIP: wait for service to be Connected
Some of our DUTs have two Ethernet devices (builtin and USB), and since
builtin Ethernet tends to get picked up first, we pick that service to
configure, even though USB is the device we have connectivity on. After
the builtin Ethernet is configured with static IP, the USB ethernet
connects and the service sort order changes, causing the static IP to be
applied to the USB ethernet. However, since the test is still under the
impression that the builtin Ethernet has the static IP, we look for it
on the wrong device and fail the test.

By waiting for the service to be Connected first, we guarantee that the
service order won't change and that we configure the correct Service.

BUG=b:161877058
TEST=run the test 20ish times on an affected device

Change-Id: I95b979018ecd85a9eacd10331b3aeddc463589b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2328134
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>