ui.WindowCycle: add sleeps between other keypresses
I recently added a delay between pressing 'Alt' and 'Tab', in order to
make sure the cycle menu appeared in the UI. It seems to have been
effective and the test is much less flaky now. However we still see a
few failures per build due to the cycle menu not appearing, so I want to
see if adding small delays after the other key presses will help clear
up the issue altogether.

BUG=chromium:1081688
TEST=tast run <ip> ui.WindowCycle

Change-Id: I5169b43c97061e07f7a0802216c23f563adf3d13