Add RC07 Tast-test for WM ARC++ R
Implement RC07: resizable/clamshell: immerse via API from maximized
portrait (pillarbox). In RC07 a resizable portrait activity is
started and it goes to maximize mode. Then the immersive button is
clicked and window bounds are compared to make sure it is in
fullscreen mode. Then the normal button is clicked, this should take
the activity back to maximize mode.
BUG=b:140203139
TEST=tast run $DUT arc.WMResizableClamshell

Change-Id: I94ae1786be3041f2b87fbaeac7c460e2f7511cbb