tast-tests: platform.BootPerf: fix test flakes
This fixes flakes in test platform.BootPerf:
* Don't check the /home/chronos/.oobe_completed after Chrome login: the
test checks whether /home/chronos/.oobe_complete is present to make sure
we boot to the user login screen instead of OOBE. If the file is not
present, we perform a Chrome login and then check the file again. The
file won't show up after Chrome login, which fails the test. Fix by
removing the check and trust the result of successful Chrome login.
* Increase the timeout for waiting until the DUT is booted: the initial
timeout of 30 seconds turns out to be too short and causes failures in
waiting for the stop of bootchart, which happens after 10 seconds on
"system-services" is started. Increase the timeout to 60 seconds to
reduce the flake.

BUG=chromium:993256
TEST=CQ dry runs.

Change-Id: Ifa100f15f86eff191d07222238dae58b5e84512b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2331399
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>