tast-tests: wifi: Use precondition in SimpleConnect.
Current tests with TestFixturePre[WithCapture] seem to work well. Also
use it in SimpleConnect, the largest WiFi test where precondition can
save the most time.

It is also more probable to have problem if we fail to properly clean
up between tests. Add precondition only in SimpleConnect tests in the
CL for ease of revert if it is causing some problems.

BUG=chromium:1034875
TEST=tast run $DUT wifi.*

Change-Id: Idd5794d64e55f420b37cc58f56419182fa77fc00