tast-tests: arc.Print: Click menu item
The destination list in print preview no longer has an expanded state.
Instead of checking for this state to determine when the destination
list has expanded, simply search for the See more... menu item in the
expanded list.

Also, click the See more... menu item instead of using the keyboard to
select it.

BUG=b:162604120
TEST=arc.Print

Change-Id: Ie824002f43a0f72f05aba6f156a6695dc0334271