firmware: Refactor RebootToMode as a Helper method
RebootToMode interacts with the servo, the RPC client, and soon also the
firmware.Config. While turning off and on, it is helpful to be able to
reconnect to the RPC client. By refactoring to be a method on the
firmware.Helper, we remove the need to set up these structures in
advance, and we allow RebootToMode to reconnect to the RPC after
rebooting.

Look how much shorter the BootMode test is!  It's incredible.

BUG=b:162068479
BUG=b:155425293
TEST=firmware.BlockingSync
TEST=firmware.BootMode.normal
TEST=firmware.BootMode.rec
TEST=firmware.Config

Change-Id: Ib8720a413f271c7b9fb770bc24ed0f7d0bfe1ca1