firmware: Create ModeSwitcher object using Helper
RebootToMode interacts with the servo, the RPC client, and soon also the
firmware.Config. While turning off and on, it is helpful to be able to
reconnect to the RPC client. By refactoring to use firmware.Helper, we
remove the need to set up these structures in advance, and we allow
RebootToMode to reconnect to the RPC after rebooting.

This CL also creates a ModeSwitcher item which tracks the Helper.

Look how much shorter the BootMode test is!  It's incredible.

BUG=b:162068479
BUG=b:155425293
TEST=firmware.BlockingSync
TEST=firmware.BootMode.normal
TEST=firmware.BootMode.rec
TEST=firmware.Config

Change-Id: Ib8720a413f271c7b9fb770bc24ed0f7d0bfe1ca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2331854
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Andrew Luo <aluo@chromium.org>