Fix arc.WMResizableConversion
I've noticed that arc.WMResizableConversion is failing on
eve-arc-r. Through the debugging, I've realized that the reason
of failure is that the window is not yet in the normal state.
ash.WaitForCondition to wait for the window to be the desired
state will fix the problem.

BUG=none
TEST=tast run arc.WMResizableConversion on eve-arc-r

Change-Id: I98ef86f0a3413e3dfb7247fe105f96af741e4d32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2333812
Reviewed-by: Armen Keshishian <armenk@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>