tast-tests: Parameterize power tests using hwdep.[No]ForceDischarge
On some platforms, force discharging a battery with ectool is not
supported. Depending on whether this functionality is supported
or not, it should be decided to use metrics from battery.
This CL parameterizes power tests to decide whether battery
metrics should be collected on a device or not. It also prevents
calling ectool on devices that do not support force discharge,
which leads the test to fail.

BUG=b:157512297
TEST=PASSed affected tests on devices without batteries

Change-Id: I9cf03fba0c7d61d4e59db8060200e013854f1aa6