tast-tests/webrtc: also poll for encoder implementation name
webrtc.RTCPeerConnnection.*_cam verify that the hardware accelerator
is used when encoding the camera feed. To do that, it reads the
"encoderImplementation" entry from WebRTC's JS getStats(). Creating
and initializing and external encoder is done asynchronouysly and
while that is happening, the webrtc code fills in a temporary
"ExternalEncoder" name. This CL changes the code to keep polling
if and while we read that string.

BUG=b:162764016
TEST=tast run DUT webrtc.RTCPeerConnection.*cam on kohaku

Change-Id: I5cdb4d9f774685ec3b24d699c522912b8c89d47d