EthernetStaticIP: wait longer for service Connected State
With the addition of CL:2328134, we now wait for DHCP negotiation to
finish on the EthernetService. However, on some devices, we consistently
fail the first DHCP request and end up retrying after a wait period.
This whole process often ends up exceeding the 8 second timeout we
currently have, so extend it to 15 seconds.

BUG=b:162796466
TEST=run on affected device

Change-Id: If5cd3918f3875dc86d62d133ea0ab23df042590d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2333814
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>