Gzip compression of trace data file
I've noticed that trace data can be large but can be compressed
through gzip (Golang supports it and Perfetto can load the gzipped
data). This CL makes the compression of the trace data in the
test, which would be better for storage use and load of copying
the data files.

BUG=none
TEST=manually

Change-Id: I150490d06d0065a12036ba06906fc3df55492bdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2337213
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>