video.Contents: Add tests for composited video.
This CL splits the video.Contents variations into two groups: *_hw that
only run on devices where NV12 hardware overlays are supported, and
*_composited_hw that run everywhere and force hardware overlays off.

The reasoning is that the visible rectangle (a common source of
problems) is dealt with differently in those two paths.

Note that just because hardware overlays are not forced off, it doesn't
mean they will be used. For example, on grunt, the h264_360p_hw video
doesn't get promoted to an overlay. Dealing with these cases will be
left for later.

BUG=b:159158467
TEST=video.Contents.h264_360p[_composited]_hw on nami behave differently

Change-Id: I7a41cd06f614df193e81da220653b0a041f0b99d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2337426
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>