Invoke disableAutomation explicitly for ui.SplitViewResizePerf
ui.SplitViewResizePerf.clamshell_mode experiences bad performance
numbers on some low-end devices. One factor of this problem seems
the accessibility (automation) feature itself, which is used to
access to the virtual desks view.

This CL explicitly invokes disableAutomation function to suppress
the effect. This has a huge side effect but I think it is okay.

BUG=chromium:1111137
TEST=ui.SplitViewResizePerf.clamshell_mode on robo360

Change-Id: Ia7a0f0ce83ea5fa25d50c07ca74d832b35907696