Wait for cycle menu to be closed in ui.WindowCycle after releasing Alt
The test is currently flaky due to the cycle menu sometimes not
appearing. It's not clear why from the screenshots and logs of failed
tests, but one possibility is that the cycle menu is not closing
immediately between iterations of the loop. Let's try explicitly waiting
for the cycle menu to close after we finish each Alt+Tab cycle and see
if the situation improves.

BUG=chromium:1081688
TEST=tast run <ip> ui.WindowCycle

Change-Id: I3fc7078588f5521734f27bcdafda422081928c3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2337226
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>