tast-tests: update repeat-test.sh to count failures.
This is just to improve some of the logging output of repeat-test.sh
It is convenient to see the number of times a failure occurs when
tracking flakiness. This also logs what iteration a failure occurs on
if using -s.

BUG=None
TEST=run repeat-test.sh forcing failures

Change-Id: I00ae7640a91539754e9871918349684f5fffacdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2340482
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>
Auto-Submit: Brendan Hansknecht <bhansknecht@chromium.org>