tast-tests: Resize PIP to minimum in video.ChromePIPEnergyAndPower.small
video.ChromePIPEnergyAndPower.small has assumed that the PIP window is
created as small as possible based on the minimum 260x146 and the video
aspect ratio 4x3, meaning it should be 260x195. Stainless shows many
failures where the test reports 274x206 instead. Now the test shall drag
the resize handle to ensure that the PIP window is as small as possible.

BUG=chromium:1093933
TEST=video.ChromePIPEnergyAndPower.small

Change-Id: Id16243d01be7c4f85fbda0980a5be0bafb65c911
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2341097
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>