firmware.Helper: Pass configDataDir into NewHelper
Previously, ConfigDataDir was manually set after creating a Helper:
    h := NewHelper(ctx, s.RPCHint(), s.RequiredVar("servo"))
    defer h.Close(ctx)
    h.ConfigDataDir = s.DataPath("firmware.ConfigDir")

But, this is inconsistent with the way that other state-derived
variables are set (rpcHint and servoHostPort).

This CL makes configDataDir into a private field of Helper, which is set
during NewHelper, just like rpcHint and servoHostPort.
This is also consistent with the design that was agreed upon in
go/tast-fw-helper.

BUG=b:162068479
TEST=firmware.BlockingSync
TEST=firmware.BootMode.normal
TEST=firmware.Config

Change-Id: I1336182a1fd6599ca8885199dabd852aebea5b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2343174
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>