ui.SettingsToggleAdvanced: fix flakiness calling focus()
The test is mainly failing when trying to focus on the 'Advanced'
button. This seems to be a timing issue where the Settings app is not
fully loaded yet. Adding a check for document.ReadyState == complete
before doing anything with the Settings app fixes the issue on heavily
failing boards (i.e. octopus/droid and hatch/jinlon with have 100% fail
rates).

BUG=chromium:1085084
TEST=tast run <ip> ui.SettingsToggleAdvanced on Droid and Jinlon, which
were failing 100% before the fix
./repeat-test.sh -r 1000 <garg> ui.SettingsToggleAdvanced

Change-Id: Ic0ac3887b7e137de5a0b7c1ee60d493cf56d209d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2343995
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>