Add RC08 Tast-test for WM ARC++ R
Implement RC08: resizable/clamshell: immerse via API from maximized
non-portrait. In RC07 a resizable portrait activity is started and it
goes to maximize mode. Then the immersive button is clicked and
window bounds are compared to make sure it is in fullscreen mode.
Then the normal button is clicked, this should take the activity back
to maximize mode.
BUG=b:140203139
TEST=tast run $DUT arc.WMResizableClamshell

Change-Id: I1c9d8aafc5174ea0053042c9f190314098a5400b