crostini: Generate fontconfig cache before running copy/paste apps
The first time the wayland backend is used for the copy/paste apps, it
generates the fontconfig cache (because it's empty), which can take ~2
minutes on a lower-end device. This will cause the copy/paste app step
to timeout and fail the test. Explicitly generate the cache before
running the apps to prevent the apps from timing out.

An alternative approach would be to simply increase the app timeout.
However, that hides the root cause. Making this a separate step makes it
clear that generating the fonctconfig cache is the cause of the slow
execution.

BUG=b:162804470,chromium:1101886
TEST=Against an elm (mt8173) device:
     tast run <DUT> crostini.CopyPaste.x11_to_wayland_download_stretch

Change-Id: I325cdc4c26b9a929894c3162c4bf9250fe3a2277
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2345891
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Anand K Mistry <amistry@chromium.org>
Commit-Queue: Anand K Mistry <amistry@chromium.org>