WIP tast-tests: Add coexistence persistence tests
These tests will verify that wifi/bluetooth operations are not affected
when the other wireless technology is disabled on reboot.

This cl uses the ui to trigger the bluetooth toggling. We don't want to
ultimately use the ui to actuate bluetooth, but I think it's useful to
keep a record of being able to turn bluetooth off this way.

BUG=b:161902990
TEST=test_that ${DUT} wifi.PersistenceWifiSansBluetooth,
wifi.PersistenceBluetoothSansWifi(board: octopus)

Cq-Depend: chrome-internal:3205694
Change-Id: Idd2f5d6047dad9ffc1fa882f4247dc41269e9440