tast-tests: Make FilesApp.RenameFile take a keyboard from the caller
input.Keyboard introduces a large delay on tablet devices (see
crbug.com/997564 for an example of this issue) and therefore the
object should be passed around and reused as much as possible.

Crostini tests already do this in the precondition, so we can pass
this down to the Files app package without issue.

BUG=chromium:1101583
TEST=tast run $IP_ADDR crostini.HomeDirectoryRenameFile.download_buster

Change-Id: If5e05bd60fe7154a6981728c0bf1302953ae3dd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2351346
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>