camera: Distinguish cases that has no camera_characteristics.conf
There are three cases that a device might not have
camera_characteristics.conf:
(1) It has USB built-in camera(s) but it does not have ARC.
(2) It only has MIPI built-in camera(s).
(3) It has no built-in camera on it. (e.g. Chromebox)

We should only run LIMITED level tests for (2).

BUG=b:163988719
TEST=tast run [DUT] camera.HAL3Recording (on soraka)

Change-Id: I59cee6eda1c2b0d6034aba55623f395a6f521a29