tast: Wait cpu.shares in arc.Cgroups.
Currently, at the end of arc.Cgroups test scenario,
it checks cpu.shares just after waiting for minimizing window.
However, it is made via D-Bus invocation triggered by minimizing window,
there's small amount of chance that the cpu shared is not yet updated.

This CL waits for the condition to stabilize the test.

BUG=b:163116372
TEST=Ran "test run ${IP} arc.Cgroups" locally.

Change-Id: I483f7277aee9b4dfbeac5a0c332754cdca495189