Add RC09 Tast-test for WM ARC++ R
Implement RC09: resizable/clamshell: new activity follows root
activity. In RC09 a resizable activity is started and a new activity
is launched and window bounds are stored. Then a new child activity
is launched by clicking on the Launch Activity button. The window
bounds of child activity  should be equal to the root activity window
bounds.
BUG=b:140203139
TEST=tast run $DUT arc.WMResizableClamshell

Change-Id: I489c275f94a21c9a061a8a55f6fad019d679583f