KernelCrash: Check for non-zero signature
We should only have a full zero signature if the stack walker logic
can't figure out what symbols are in the kernel stack. Let's check to
make sure that it isn't all zero so that we can catch breakage in the
stack walker logic if the kernel decides to change it again.

BUG=None
TEST=tast run $DUT  crash.KernelCrash.* passes

Cq-Depend: chromium:2354731
Change-Id: I98114f52864267166603d725a923e7b7d5322028