arc: New AnimationPerformance app for performance metrics testing
ArcAnimationPerformanceTest app renders a GridLayout with 25x15
colored views. App shuffles the order of these views periodically
forcing a rerender of all the views. This can be arbitrarily heavy
as a rendering load based on the frequency of shuffles and the size
of the grid. App will be used for performance testing.

BUG=None
TEST=Locally integrating into arc.RuntimePerf

Change-Id: I5bb8756ec4edaa545f92a4e64c47a997477ae059