Fix ZipPerf failures
An important part of the failures for the ZipPerf tast tests are
following one of the 2 scenarios:
1. After trying to right-click and click 'Zip selection', the delete
button of the context menu is clicked instead, as shown on the
screenshot taken after a test fails.
2. After trying to name and create a new folder, the tests fails
with the error messages "Opening various_documents failed" or
"Opening 500_small_files folder failed".

The first error is related to b/161438238.

The second error is likely to be caused by two successive 'Enter'
events being registered too quickly. Now waiting for the rename input
field to disappear before sending the second 'Enter' event.

BUG=chromium:407010
TEST=tast run DUT_IP filemanager.ZipPerf

Change-Id: I3e5c44046a97199aba9fb805550e2896d5f05ee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2366519
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Tested-by: Jeremie Boulic <jboulic@chromium.org>
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>