tast-tests: Add cryptohome.RegularSession
We'll be changing how regular user sessions are mounted,
so add a test that we can use during development. The changes in the
regular session mount/unmount will be activated behind a compile time
flag with the below mentioned CL. This test will help during the work
to remove the flag. .

BUG=chromium:1074735
TEST=Build and deploy cryptohome with and without USE='mount_oop' with
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2321614
TEST=tast run cryptohome.RegularSession passes.

Change-Id: I604fcdd4d94de1c84122a26c854a1f1fcec09aac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2363093
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>