tast-tests/webrtc: rename MediaRecorder tests
MediaRecorder and MediaRecorderAccelerator are incorrectly
named: the former should be MediaRecorderAPI and the latter
should take over as simply MediaRecorder, to be homogeneous
with the other video.* and webrtc.* test cases.

Renaming is somewhat traumatic because it prevent bisects
and makes test monitoring less effective, so I waited until
there's long stable streak, and we're about to branch, to
post this CL.

I took the chance to reduce the frequency of some of these
tests to graphics_nightly ISO graphics_perbuild, since they
are not so critical.

BUG=chromium:1102417
TEST=tast run DUT webrtc.MediaRecorder.* webrtc.MediaRecoderAPI
on kohaku

Change-Id: Ifaa108ffcc58c641da3fc26bef22f49790f2aa05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2363177
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Miguel Casas <mcasas@chromium.org>