tast-tests: platform.BootPerf: use a shorter, 30 sec for CPU cooldown
Use a timeout of 30 seconds for cooling the CPU to avoid the timeout that
each iteraion takes too much time (5 mins) cooling the CPU and then the
test can't finish running the default 10 iterations within 15 mins. The
timeout of 30 seconds comes from the observation in test logs that the
CPU temperature drop is marginal after 30 seconds.

BUG=chromium:1119302
TEST=tast run $IP_ADDR platform.BootPerf

Change-Id: I2bc1fac7eae526ab558543ceb6e22107a116b64e