Add NC15 Tast-test for WM ARC++ R
Implement NC15: non-resizable/clamshell: display size change.
In NC15 set of non-resizable activities are started and button bounds
are stored. Then a zoom factor is changed and button bounds after
zoom is stored, because the window is maximized, the content will be
relayout, thus the button bounds must be different. DPI changes
proportionally must be equal to the zoom factor.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableClamshell

Change-Id: I494771d1b9800dc24469e4c09a3a6b45b1ffd8db