trace_replay: Update result report protocol
Change the existing protocol between the host and the guest part of
trace_replay so that the type and the content of test results are
decided by the guest. It should be possible to define different result
sets for individual test groups and type of tests.

BUG=b:165737380
TEST=tast run --buildbundle=crosint ${DUT}
graphics.TraceReplay.all_short

Cq-Depend: chromium:2366063
Change-Id: Iac0b366d1ea13b4004618301d4c95ddd77984303