trace_replay: Update result report protocol
Change the existing protocol between the host and the guest part of
trace_replay so that the type and the content of test results are
decided by the guest. It should be possible to define different result
sets for individual test groups and type of tests.

BUG=b:165737380
TEST=tast run --buildbundle=crosint ${DUT}
graphics.TraceReplay.all_short

Cq-Depend: chromium:2366064
Change-Id: Iac0b366d1ea13b4004618301d4c95ddd77984303
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2366672
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Robert Tarasov <tutankhamen@chromium.org>
Commit-Queue: Robert Tarasov <tutankhamen@chromium.org>