Add RT12 Tast-test for WM ARC++ R
Implement RT12: resizable/tablet: hide Shelf.
In RT12 two sets of activities are tested. A portrait + unspecified
and landscape + unspecified. Each activity starts in tablet mode and
depends on its orientation, the display should rotate to match with
the desired orientation of the activity. Then the shelf will be
hidden and shown and at each state the window bounds are stored to be
asserted at the end.
Unspecified activities are running on the same orientation as the
previous activity, so in case of portrait + unspecified, the
unspecified desired orientation is portrait.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableTablet

Change-Id: I91fcf75f2f2696ac6f3b78261faaf9952bd1c26e