tast-tests: Add USF test for non-maxized (normal) state.
Ensure that USF (uniform scale factor), is behaving correctly for
non-maximized ARC++ applications.
The application is first launched, to obtain a baseline for the number
of black pixels. This value is then used to verify the state of the
launched application, after enabling USF.

This CL also does some refactoring, to clean up common logic among USF
tests.

BUG=b:163656122, b:164507920
TEST=tast run DUT_IP arc.UniformScaleFactorNonMaximized

Change-Id: I03f542413a12df6c0b35c973d0ae9c436e474d2e