arc: Wait ARC is stopped in auth_perf.
This fixes races in arc.AuthPerf* tests. In container we rely on fact
that data removed is signal that ARC stopped. However this no longer for
ARCVM which requires 6+ seconds to stop. This problem normally was
hidden in loop wait for CPU cool down. During that time ARC was usually
stopped. In case of failure timeout happens and CPU usually is already
cooled down and we trying to start next ARC and waiting previous
instance is stopped. This leads to significant flakiness in test
results.

BUG=b:165934650
TEST=Locally, usual range of samples is much narrowed compared case
before the fix, 8sec min/max -> 5sec.

Change-Id: I0361acfc27aa4601d3fa0db7b1722454b6890c2b