arc: Wait ARC is stopped in auth_perf
This fixes races in arc.AuthPerf* tests. In container we rely on fact
that data removed is signal that ARC stopped. However this no longer for
ARCVM which requires 6+ seconds to stop. This problem normally was
hidden in loop wait for CPU cool down. During that time ARC was usually
stopped. In case of failure timeout happens and CPU usually is already
cooled down and we trying to start next ARC and waiting previous
instance is stopeed. This leads to significant flakiness in test
results.

BUG=b:165934650
TEST=Locally, usuall range of samples is much narrowed compare before
fix.

Change-Id: I0361acfc27aa4601d3fa0db7b1722454b6890c2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2369674
Tested-by: Yury Khmel <khmel@google.com>
Commit-Queue: Yury Khmel <khmel@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Auto-Submit: Yury Khmel <khmel@google.com>