crostini: Auto-generate crostini test parameters
Crostini tests currently need to have many different parameterized
sub-tests to cover all the different ways crostini can be downloaded
and installed. We want to generate them dynamically instead.

This CL migrates most crostini tests to use the new generated test
parameters. The remaining tests are special cases which will be
handled in separate CLs.

*** Why this is needed, despite being a style violation ***
We are planning to change the way crostini is deployed to end users
from using component updater to the new DLC service. This means that,
at a minimum, we are going to have to support tests which use both
mechanisms until we fully switch over. Therefore, at a bare minimum,
we need to add "dlc" and "dlc_unstable" test versions to the current
collection of "artifact", "artifact_unstable", "download_stretch", and
"download_buster".

But this minimum shift leaves us with other issues:

Firstly, only a single container version can be tested on the CQ, as
download tests are intrinsically flaky. This might not be considered a
huge issue as we are just about to finish the transition from stretch
to buster anyway, but bullseye (Debian 11) is expected to be released
in about 4 months and we will have the same need to test both shipping
versions as we currently do. Note that there was a recent regression
on stretch (issue 1120277) that was only caught by manually checking
informational tests, so this lack of testing has produced real issues.

Secondly, the current way we deploy containers to artifact tests (as
build artifacts from the build-bots) makes it impossible for external
parties to run these tests, as they don't have access to the relevant
GCS buckets. It's also difficult to run tests on local builds for
similar reasons.

Third, build artifacts force a copy of the container to be stored for
every build of every board, even though the data is mostly
unchanged. This puts unnecessary load on cros build infra.

We could fix the first issue by making stretch and buster (or buster
and bullseye) build artifacts, but this would exacerbate the third
issue. We would also have to keep around the existing download tests
for a total of 10 test versions (up from 4):
{stretch,buster} x {stable,unstable} x {component,dlc} + stretch and
buster download tests.

We could also install the container as a DLC which is created solely
for testing purposes. This would let us remove download tests (as
external parties could run the dlc tests) and fix the first issue, but
still exacerbates the third. This leaves us with 8 versions of each
test:
{stretch,buster} x {stable,unstable} x {component,dlc}

Finally, we could deploy the container as a regular external data
dependency. This still lets us get rid of download tests and test
multiple container versions on the CQ, and it means no extra copies of
the container need to be stored, but this multiplies the number of
test parameters needed even further. We actually make both amd64 and
arm versions of each container. Currently the build-bots take on the
responsibility of downloading the correct version for their board, but
without that we would need to have separate test parameters for each
architecture, for a total of 16:
{stretch,buster} x {stable,unstable} x {amd64,arm} x {component,dlc}

We prefer the final option here because it solves all of our issues,
except for the number of test parameters required.

Note in particular that this option would not increase the testing
load, because any given device would only be one of stable or
unstable, and one architecture, so only 4 kinds of test could run,
only one more then the current situation. We could reduce this further
by only running component tests on devices that don't support dlc once
we're willing to enable dlc by default on any devices, which would
reduce each device to just 2 versions of each test, better then the
current situation. Either way, all the tests running would have the
potential to be promoted to the CQ.

I claim that for all the options here, but especially the preferred
one, the number of test parameters is excessively large for manual
maintenance.

The main argument for explicitly listing test parameters every time in
writing_tests.md is that "test registration should be
declarative". This is a desirable quality, but it's much more valuable
when tests have different sets of parameters, rather then the same set
duplicated many times across many different tests. I claim that
declaring that a test uses the standard set of crostini parameters is
easier to understand, and easier to write, then listing them
explicitly every time.

Further, when a test has actual test-specific parameters, such as
copy_paste.go, the inflated list of general crostini parameters
obscures what is special about this test. It is much more
understandable to list the test-specific parameterization and generate
the full set from that.

BUG=chromium:953544
TEST=tast run $IP_ADDR '("group:mainline" && !informational)'
TEST=tast run $IP_ADDR crostini.*
TEST=In each case, check that the correct list of tests ran

Change-Id: I2df897c6a7e808e8d3494e4d5054565e244a7a9a